Adding DataDog Sink
Adding a sink for DataDog. Updating `MetricValue` to include a `Frequency` field which records the frequency the metric is collected at. This value is need by DataDog's `COUNT` metric type in order to display metrics properly in the UI at various resolutions.

Added `collectorFrequencyWrapper` type to wrap collectors and update the `Frequency` field on the returned `MetricValue`s.