Add monitor reloading
Adding code to reload monitors via the API. Per comments in `loader.go` above `StartMonitors` it _seems_ like this functionality should exist. `loader.Load` supports detecting diffs in monitors but there is no hook from the API to run `Load` then `StartMonitors`.

Added a `Stop` function to `Engine` to facilitate cleaning up metrics when a monitor stops. The level collector has code but no cleanup on the engine, and the metrics it contains, is ever run. This leaves a heartbeat reader hanging for old monitors using the `repl.lag` metric.