Hotswappable Credentials
Added support for full credentials (username, password) to DSN reload for driver. In some cases we have the contents of a CNF file updated with both new credentials and a new user name, and possibly updated certs. This supports that case while preserving the original functionality around a fixed username and rotated passwords. TLS reloading is handled inside the `CredentialsFunc` and is transparent to the DSN hotswap.

This doest _not_ support turning TLS on/off dynamically. This approach assumes that once we choose to use TLS for a monitor that we will continue to do so, and we will only rotate the certs. TLS reloading is only run if TLS was originally configured.