Moving Package Variable to Monitor
Moved the package-level variables `statusQuery` and `newTerms` to the monitor in case multiple monitors within the same domain are created but are running on different MySQL versions.

Added a check around `replStatus` when getting the "source" entry. In cases where something is not a replica the map will be nil and should not be accessed.