Fixing intermittent panic on metric submission
In certain (unknown) cases, submitting metrics via the Datadog API fails with the error `EOF` and a `nil` http response. It appears that while the API config has a `compress` field, setting it to `true` does not actually enable compression. It only prevents it from disabling compression:
```golang
if !c.Cfg.Compress {
		// gzip is on by default, so disable it by setting encoding to identity
		localVarRequest.Header.Add("Accept-Encoding", "identity")
	}
```

Despite the Datadog client indicating the gzip compression is on by default, an inspection of the code and dumping the request indicates that it is not. Manually setting the content encoding on each request is required to actually enable compression.

In some case the lack of compression caused a failure in the underlying http client which resulted in an `EOF` error. It's unclear what payload size or conditions resulted in this issue, but adding the content encoding resolved the issue.

- Added content encoding of `gzip` when `compress` is set to `true`
- Added a `nil` check on the http response to prevent trying to get the status code from it when it is `nil`