Truncation Timeouts
Added timeouts on truncation operations for `query.response-time` and `wait.io.table`. In MySQL [locking an instance for backup](https://dev.mysql.com/doc/refman/8.0/en/lock-instance-for-backup.html) will cause `TRUNCATE` to block. This will cause multiple instances of the command to stack up, driving up the thread count on the server. A default timeout of `2s` has been set and is configurable by the user. An error, `truncate-failed` has also been setup to allow users to specify an error handling strategy in the event that `TRUNCATE` fails.