`lock_wait_timeout`
Adding a call to set `lock_wait_timeout` on the session in order to try and prevent truncate from running too long. `context.Timeout` doesn't cause a blocked `TRUNCATE` query to stop when the timeout occurs so we need to leverage setting `lock_wait_timeout` on the session to prevent the statement from blocking for a long time.