Adding `DeltaSink`
Added a new sink, `DeltaSink`, which performs the calculations necessary to convert `CUMULATIVE_COUNTER` values into `DELTA_COUNTER` values. This is necessary as converting counters to `DELTA` counters needs to happen _before_ values are sent to a `RetrySink`. Due to how the `RetrySink` is constructed, errors. during metric can result in the `RetrySink` sending batches of metrics out of order. Out of order submission of metrics to sinks will cause unexpected values to be calculated for delta counters for those sinks that internally perform delta calculations (like the Datadog sink) as the sinks expect values to always be sent in-order so that the delta logic works.

The `DeltaSink` perform the calculations internally and substitutes calculated `DELTA_COUNTER` values for any `CUMULATIVE_COUNTER` values and then sends them to the sink that it wraps. If the wrapped sink is a `RetrySink`, and error during submission within the `RetrySink` do not cause issues as the delta calculations have already been performed based on the in-order metrics and cached in the metric batch sent to the `RetrySink`.

This includes several unit tests in `datadog_test` that demonstrate the behavior of the `RetrySink` issue with delta calculations, as well as an example of the `DeltaSink` resolving the problem.