Force Order for Query Latency
`query.response-time` relies on the ordering of results based on `bucket_quantile` but in some cases there are multiple rows with the same value for `bucket_quantile`. This results in undefined behavior as we don't know which of the rows returned will be chosen for the `LIMIT 1`. Usually the results are ordered by the natural order of the table (e.g. `bucket_number`) which produces expected behavior but occasionaly this isn't the case and anomalous values are returned as the latency. Adding an explicit ordering on `bucket_number` should resolve this issue.